content [nfc]: Follow browser preference setting when opening links
This preserves the previous behavior (external application on iOS,
and platform default on Android) when browserPreference is `null`.

There is no way for the client to update the setting for now, hence the
NFC.

The helper could have been a static method
BrowserPreference.getUrlLaunchMode, similar to ThemeSetting.displayName,
but that's a lot more verbose for the caller.

This also updates references to use the alias `UrlLaunchMode` in the
comment instead, since that's the type the developer sees in this
context.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>