settings: Add browser preference setting ui
The translation for the label is taken from zulip-mobile. The user
facing string refers to "in in-app browser" when the actual value is
`UrlLaunchMode.platformDefault`. This is consistent with the behavior
the user sees when the link is a HTTP URL on Android and iOS, which
should cover the majority of use cases of this setting (and probably
what the user would expect).

The UI currently does not have a design, and is made to be as simple as
possible to implement for now.

Fixes: #1228

Signed-off-by: Zixuan James Li <zixuan@zulip.com>