compose: Change content input hint text if topic is empty and mandatory
Previously, "Message #stream > (no topic)" would appear as the hint text
when the topic input is empty but mandatory.  Now, it is shown as
"Message #stream" instead, since the "(no topic)" isn't allowed when
topics are mandatory.

The control flow of `getDestinationString` can be simplified. However,
it is structured this way to prepare for a later change to support
showing "general chat" in the hint text, that adds some more advanced
checks.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>