compose [nfc]: Handle empty topics for topic-narrow input hint text
While this appears to be a user facing change, it's not visible yet, not
until TopicName.displayName becomes nullable.  This is a part of a
series of changes to handle empty topics.

A test is skipped because the server does not send empty topics to the
client without "empty_topic_name" client capability.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>