store [nfc]: Use try/finally to clear _perAccountStoresLoading[accountId]
In GlobalStore.perAccount, the loadPerAccount future can throw, and
in that case the `_perAccountStoresLoading.remove` in these lines of
code hasn't been getting called.

That's only a latent bug, currently, because loadPerAccount only
throws if the account was logged out, and we always do a
`_perAccountStoresLoading.remove` on logout (in
GlobalStore.removeAccount).

But we'd like to add another case where the loadPerAccount future
throws (disallow connecting to ancient servers), which doesn't
involve logout / removeAccount. So this try/finally will be
important for that case.