content test [nfc]: Use checks-extensions in list-marker tests; explain a bit more
A couple of the extension members added here don't actually end up
getting used (because the "height of one line" check feels like it
benefits from describing the calculation more explicitly).  Still
the fact we're referring to those properties in a test makes a good
prompt to add the extension so future tests can use it.