message: Enhance message move checks
Because of this simplification, it is easier to make sense of the
constraints that apply to the inter-related fields.

This also allows us to drop the `assert`'s and "TODO(log)"'s, because
the stacktrace we need can be retrieved after throwing these
`FormatException`'s.

This is similar to zulip-mobile code for parsing move data. The main
difference is that we now also check the value of `propagate_mode`,
which is documented to be present on message moves.
See Greg's comment on this:
  https://github.com/zulip/zulip-flutter/pull/1311#issuecomment-2667327271

Signed-off-by: Zixuan James Li <zixuan@zulip.com>