logo: Make source SVGs self-contained
This makes these SVGs a bit easier to reason about.  More
concretely, this will allow us to switch from Inkscape to librsvg
for the conversion: the latter is faster and has some other useful
properties, but doesn't support using xlink to point to a separate
file as these were doing.

This change is nearly NFC: it'd be NFC if the image processing were
mathematically exact.  In reality the old and new output images look
identical to my eye, but examining with `sng` (like in the previous
commit upgrading Inkscape) confirms some pixels have slightly
different values.  Like with that commit, we leave those changes out
for now, to bundle them with substantive changes coming later in
this commit series.