setting [nfc]: Make in-app explicitly the browser preference on Android
The url_launcher documentation warns that the meaning of
platformDefault may change at any time.  We've chosen
particular behaviors we want on different platforms, so
express those choices explicitly.

This change is NFC given the current behavior of url_launcher,
which appears to be that platformDefault means an in-app browser
for at least HTTP links.

Co-authored-by: Greg Price <greg@zulip.com>
Signed-off-by: Zixuan James Li <zixuan@zulip.com>