settings [nfc]: Split out a GlobalSettingsStore, and GlobalStoreWidget.settingsOf
This allows widgets to listen separately for changes to settings or
to the rest of the global store's data, which is a bit nice.

More fundamentally, as a matter of code organization, it gives us a
place to put code related to the store of global settings where
there's more room to spread out than if we kept it in GlobalStore,
and where the namespace is already specific to global settings.

Those two points in turn will also help us adjust the API the rest
of the app uses for interacting with these settings, so as to make
it more ergonomic.

The TODO comments added in this commit mark places where this API
will be more coherent after upcoming commits in this series.