content [nfc]: Inline showError helper in _launchUrl
Some of _launchUrl is specific to opening links in message content,
but a lot is generic and could helpfully be pulled out to be reused
in other places where we open links. We'll do that next; first,
inline this helper function, which has one caller in the
content-specific code and one in the generic code.