notif [nfc]: Change `_groupKey` method to take a realm URL and user ID
The `_groupKey` method currently takes an `FcmMessageWithIdentity`
object as an argument. However, this method only requires the
realm URL and user ID from that object. This change modifies the
method signature to directly accept `realmUrl` and `userId` as
arguments.

This change is made to facilitate calling `_groupKey` from
contexts where an `FcmMessageWithIdentity` object is not
available.

Discussion:
  https://chat.zulip.org/#narrow/channel/516-mobile-dev-help/topic/Removing.20notifications.20for.20logged.20out.20account/near/2088813