recent_senders: Handle moved messages
The tests do not have to use PerAccountStore, but this setup makes it a
bit more integrated.

This is similar to how we add support to handling moves for unreads
(commit 34e6201f), also with optimizations to avoid making unnecessary
copies of message IDs when the entire conversation is moved (e.g.
resolve/unresolve topic).

An alternative approach to this that we didn't take is extracting
helpers from handleMessages and handleDeleteMessageEvent and combining
the two to handle moves, like web does
(https://github.com/zulip/zulip/blob/bd04a30b/web/src/recent_senders.ts#L165-L190),
but creating a dedicated helper makes it more sraightforward to
optimize for our performance need.

Fixes: #901