db [nfc]: Treat schema version as a static const, not instance method
Conceptually this is really a static fact about the class, not a fact
about a given instance of it.  (The implementation's body is always
just a constant integer.)  Make things a bit clearer -- and make it
easier to get hold of the value from outside the class -- by making
that explicit, with a static const field as the home of the value.