wip scroll: Keep short threads pinned at bottom, not past bottom; TODO test
This is NFC as to the real message list, because so far the bottom
sliver there always has height 0.

This is a step toward letting us move part of the message list into
the bottom sliver, because it fixes a bug that would otherwise create
in the case of a thread that fits entirely on the screen.