compose: Fix empty topics not being shown correctly in hint text
Before, if the user chooses an empty topic, the hint text
might appear as: "#issues > ".

Now, with this fix, the hint text will be
(correctly): "#issues > general chat"

This buggy logic was introduced in 769cc7df, which assumed that
TopicName.displayName is `null` when the topic is empty.
TopicName that came from the server are guaranteed to be non-empty,
but here our code can construct an empty TopicName, breaking this
assumption.

To enable this fix while the rest of the app still relies on
TopicName.displayName being non-nullable, switch to using plain
strings here for now.  Later once TopicName.displayName becomes
nullable, we'll go back to constructing a TopicName here.