compose: Unskip test; show realmEmptyDisplayName in hint text
This logic was introduced in 769cc7df, which assumed that
TopicName.displayName is `null` when the topic is empty.
TopicName that came from the server are guaranteed to be non-empty,
but here our code can construct an empty TopicName, breaking this
assumption.

Switch to using plain strings, and go back to constructing TopicName
with empty topics once TopicName.displayName becomes nullable.