wip scroll: Go to end at first, and stay at end once there; TODO split?; TODO test
This is NFC as to the real message list, because so far the bottom
sliver there always has height 0, so that maxScrollExtent is always 0.

This is a step toward letting us move part of the message list into
the bottom sliver, because it means that doing so would preserve the
list's current behavior of starting out scrolled to the end, and its
current behavior of remaining scrolled to the end once there as e.g.
new messages arrive.