message: Create an outbox message on send; manage its states
While we do create outbox messages, there are in no way user-visible
changes since the outbox messages don't end up in message list views.
We create skeletons for helpers needed from message list view, but don't
implement them yet, to make the diff smaller.

---

OutboxMessage is designed to keep the timers internal to itself, and
expose only the callback functions, to keep message store code simple.

To make sure that OutboxMessage.dispose is always called when the outbox
message is removed from the store, we expose an unmodifiable view of the
actual map.

For testing, similar to TypingNotifier.debugEnabled, we add
MessageStoreImpl.debugOutboxEnabled in case a test does not intend to
cover outbox messages.