example_data [nfc]: Add messageEvent
MessageEvent constructor calls are generally quite short, and there are
no boring required fields other than an integer id.  However, as we add
localMessageId to the event, it would be expected to have a boring
value most of the time, so prepare for that.

A neat side-effect is that we can cut some imports from lib/api/model/.

(Some places that use non-zero event IDs now use 0, but we don't
care much about this value anyway.)