api [nfc]: Pull out DisplayableMessage
See also CZO discussion on the design of this, and the drawbacks of
the alternatives:
  https://chat.zulip.org/#narrow/channel/243-mobile-team/topic/A.20new.20variant.20of.20Zulip.20messages.20-.20inheritance.20structure/with/2141288

This will help support locally echoed messages in MessageListView later.

This requiring specifying the element type of `List`s in some tests (or
in some cases, the type of a variable declared).

This is a side effect of making `StreamMessage` and `DmMessage` extend
`Message<T>` with different `T`'s. When both appear in the same `List`,
the upper bound is `Object`, instead of the more specific
`Message<MessageDestination>`.
See "least-upper-bound" tagged issues for reference:
  https://github.com/dart-lang/language/issues?q=state%3Aopen%20label%3A%22least-upper-bound%22