msglist [nfc]: Migrate most MessageListItem subclasses to use MessageBase
except MessageListMessageItem.

This keeps changes minimal, leaving most of the helpers in
lib/model/message_list.dart untouched, to avoid unnecessary
generalization.

This hoists streamId in StreamMessageRecipientHeader.build,
where we used to access streamId via message instead from the
onLongPress callback.  Because Message.streamId only changes on
message moves, we expect the build method to be called again, so
this should be fine.