content: Support basic text styles for KaTeX content
This adds another experimental flag called `forceRenderKatex`
which, if enabled, ignores any errors generated by the parser
(like when encountering an unsupported CSS class) tries to do
a "broken" render of the available span and their styles.
Allowing the developer to test the different KaTeX content in
the wild easily, while still in development.