emoji: Parse URLs the same way for validation as for use
I think this change might be NFC, but I'm not sure.  The situation
where it could change the behavior is if there's some URL string
that would be accepted by `Uri.parse`, but rejected by `Uri.resolve`
with the realm URL as base URL, or vice versa.

If there is any such situation, then this behavior is more accurate.
If there isn't, then this is just a small simplification.

Noticed this because it's one of the handful of references to
the store's realmUrl, and another one is a call to tryResolveUrl.