user [nfc]: Move selfUserId to PerAccountStoreBase
This `selfUserId` getter has 16 other references across 11 files.
Thankfully our architecture, and in particular this way of combining
the substores:

  class PerAccountStore extends PerAccountStoreBase with ChangeNotifier,
    EmojiStore, UserStore, ChannelStore, MessageStore {

has meant that those all refer to it on the main PerAccountStore type,
rather than on an individual substore like UserStore.  So none of
those need to be touched when we rearrange like this where the data
is stored.