msglist: Fix speed calculations in scroll to bottom
Fixes #1485.

This logic for deciding how long the scroll-to-bottom animation
should take, introduced in 5abeb887e (#223), didn't have any tests.

As a result, when its unstated assumption about the message list --
that scrolling up from the end was represented by positive scroll
positions -- was broken a few months later (in bdac26fb1), nothing
alerted us to that.

I did notice a few times over the past year or so that the effect of
the scroll-to-bottom button seemed jerky, as if it were trying to
move much farther in each frame than it should.  Now I know why.

(Discovered this in the course of revisiting this code in order to
adapt it to the more radical change to the message list's scroll
positions which is coming up: "zero" won't be the end, but somewhere
in the middle.)