scroll [nfc]: Move scroll-to-bottom logic into our ScrollPosition subclass
This makes a more comfortable home for extending this logic,
because it naturally uses a lot of ScrollPosition members.

To make the logic fit in better on this class, also loosen its
assumptions slightly, allowing maxScrollExtent to be nonzero.
It's still expected not to change, though -- fixing that is a more
complex job, and will come over the remainder of this commit series.