scroll: Apply a min velocity in ScrollToEndSimulation
This makes it possible to see in a self-contained way, in this class's
own code, that it always starts moving at a velocity that isn't zero,
or less than zero, or at risk of being conflated with zero.

This doesn't have a big effect in practice, because the only call site
already does something else whenever the distance to travel is
negative or very close to zero.  But there is a small range -- namely
where the distance to travel is between 1 and 12 physical pixels,
given the default behavior of ScrollPhysics.toleranceFor -- in which
this minimum speed does apply.