dialog [nfc]: showSuggestedActionDialog returns DialogStatus with result
With this API, we can use showSuggestedActionDialog in an "early
return" style -- await the user's choice, and early return if it was
"Cancel". That's particularly helpful when the confirmation dialog
belongs in an `if` block.

We'll use this for the upcoming "edit message" feature (#126), where
we plan to offer a confirmation dialog before entering an
edit-message session, but only if the compose box has text for a new
message in it (which would be discarded if the user wants to go
ahead).