compose: Change topic input hint text
This is similar to web's behavior.  When topics are not mandatory:

- an alternative hint text "Enter a topic (skip for “general chat”)"
  is shown when the topic input has focus;

- an opaque placeholder text (e.g.: "general chat") is shown if
  the user skipped to content input;

Because the topic input is always shown in a message list page channel
narrow (assuming permission to send messages), this also adds an
initial state:

- a short hint text, "Topic", is shown if the user hasn't
  interacted with topic or content inputs at all, or when the user
  unfocused topic input without moving focus to content input.

This only changes the topic input's hint text.

See CZO discussion for design details:
  https://chat.zulip.org/#narrow/channel/530-mobile-design/topic/general.20chat.20design.20.23F1297/near/2106736