compose: Change content input hint text statefully
Before, the content input, when empty, shows the
"#stream > general chat" hint text as long as it has focus,
and shows "#stream" as the hint text when it loses focus.

Now, when the content input is empty, it still shows
"#stream > general chat" when it gains focus, except that it will
keep showing it even after losing focus, until the user moves
focus to the topic input.