compose: Change content input hint text statefully
Before, the content input shows the "#stream > topic" hint text
as long as it has focus, and set the hint text to "#stream" when it
loses focus.

Now, the content input still shows "#stream > topic" when it gains
focus, except that it will keep showing it even after losing focus,
until the user moves focus to the topic input.