deps: Update file_picker to 10.1.2, from 9.2.1
Changelog:
  https://pub.dev/packages/file_picker/changelog#1012

There are two breaking changes, one about deprecating an option
`allowCompression` and changing the default to false (was true),
and another about changing default for `compressionQuality` to 0
(was 30). This means previously it used to compress media by
default and now it doesn't. See discussion for why we take in the
new defaults:
  https://github.com/zulip/zulip-flutter/pull/1474#discussion_r2048048498

There is also a lint error that is caused by a dependency of
package:file_picker. I didn't find any reports about it upstream,
and we probably aren't affected by it (confirmed by running simple
upload tests of a PDF, an XML file and an image, and didn't see any
errors/crashes), so for now added a suppression for that error.
Also we're planning to switch away from this library anyway (#855).