msglist [nfc]: s/composeBox{Controller,State}/ in MessageListPageState
Soon we'll want a method for starting a message-edit session, and
it'll make sense to put it on ComposeBoxState rather than
ComposeBoxController; one reason is that it'll need UI code to show
an error dialog.

With this change, the ComposeBoxState -- and so that new method --
will be accessible in the message-list widget code, where we open
the message action sheet from.