msglist: Drop easing curve in scroll-to-bottom
Fundamentally an easing curve like this relies on knowing in advance
how far the animation is going to end up going.  When we start
letting the message list scroll from the middle of history, it'll no
longer be possible to know that.

Switch instead to a behavior that can be based on only what's happened
so far, not on a prediction of the future.  In the future, if we want,
we could get fancy and make the speed change over time; but to start,
keep it simple, and just move at the same speed from start to finish.