katex [nfc]: Make KatexSpanStyles immutable
Once the parsing is done, we want these to remain unchanged,
just like the other objects in the parse tree.

So, like ContentNode and its subclasses, make the class immutable.
The parser needs to mutate its own draft of what styles to apply
to a given span; but it can do that with its own local variables
corresponding to the fields, and construct a styles object at the
end of the loop.