api [nfc]: Extract Conversation.isSameAs
This will make it easier to support comparing the conversations
between subclasses of MessageBase.

The message list tests on displayRecipient are now mostly exercising the
logic on Conversation.isSameAs, which makes it reasonable to move the
tests.  Keep them here for now since this logic is more relevant to
message lists than it is to the rest of the app.

Dropped the comment on _equalIdSequences since it is now private
in the short body of DmConversation.