scrolling test: Skip tests against old default z-order of slivers
These two tests on CustomPaintOrderScrollView compare its behavior
to the default z-order / paint order behavior of CustomScrollView.
They'd therefore be broken by my upstream PR that gives
CustomScrollView the functionality of CustomPaintOrderScrollView
and simplifies its default behavior:
  https://github.com/flutter/flutter/pull/164818

So skip the tests.  Once that PR lands, we can update past it and
then remove CustomPaintOrderScrollView and its tests entirely.

(The customer_tests failures currently seen on the upstream PR are
different, because the Zulip pin in flutter/tests is currently at
d302584e5.  The failures there are in the sticky_header tests, which
want to exercise a variety of sliver configurations and need to
control the z-order when they do so.  As of d302584e5, they did so
by relying on the specifics of the existing CustomScrollView
behavior.  Since then we merged CustomPaintOrderScrollView, and in
49eff8bb3 adapted those tests to use it, making them immune to
changes in the default z-order; but that also meant adding
CustomPaintOrderScrollView's own tests, including these two.)