channel: Maintain maps consistently when removing subscriptions
When a subscription is removed, the data in streams and streamsByName
are left as instances of `Subscription`.  This is inconsistent with the
data store's assumption that unsubscribed channels should have plain
`ZulipStream` instances.

This doesn't seem to have user-facing effects, though.  The assertion
errors only affect debug builds.

CZO discussion:
  https://chat.zulip.org/#narrow/channel/48-mobile/topic/Unsubscribe.20then.20resubscribe.20to.20channel/with/2160241