scroll [nfc]: Clarify scroll-to-end calculations a bit more
This version keeps the numbers in the form of doubles, with seconds as
the unit of time, until the end.  That's a bit more typical Flutter
style, and also brings it closer to how the logic will look when we
flip this around to produce a velocity instead of a duration.