msglist: Simplify computing semanticChildCount
This has already gotten out of sync with the total number of
children in the list, after 5c70c76f0 added a TypingStatusWidget
child while leaving this unchanged.

On the other hand, its spec isn't to be the total number of
children: it's to be the total number of children "that will
contribute semantic information".  So the spacer shouldn't count,
and probably TypingStatusWidget and MarkAsReadWidget shouldn't
either when they don't show anything.  (Currently it effectively
counts the spacer and MarkAsReadWidget, or one could retcon that
to TypingStatusWidget and MarkAsReadWidget.)

Further:

  > If the number is unknown or unbounded this should be left unset
  > or set to null.

So it seems like when `haveOldest` is false, this should be null.

Meanwhile, the exact value here seems to have little effect.  I just
tried the app out in TalkBack, and there are some other issues (when
you scroll around, there's some glitchy behavior with focus moving
to the sticky header and back, and even with the scroll position
sometimes abruptly jumping back; and I couldn't find a way to
navigate by a whole message or item at a time, though that might
have been my inexperience with TalkBack) but it doesn't ever quote
the number of list items as far as I could tell, and isn't bothered
by this being a few more or less than the spec'd number.

So for the moment let's do something simple: don't count any of the
extra children at the end, the ones that are often or always empty.
This also makes a couple of tests a bit simpler to follow.

Thanks to Zixuan for raising the question.  Discussion here:
  https://github.com/zulip/zulip-flutter/pull/1468#discussion_r2067407272