topics: Add topic list page
For the topic-list page app bar, we leave out the icon "chveron_down.svg"
since it's related to a new design (#1039) we haven't implemented yet.
This also why "TOPICS" is not aligned to the middle part of the app bar
on the message-list page.

We also leave out the new topic button and topic filtering, which are
out-of-scope for #1158.

The topic-list implementation is quite similar to parts of inbox page
and message-list page.  Therefore, we structure the code to make them
look similar to compare for changes side-by-side to help with reviewing
what has changed.

Figma design:
  https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=6819-35869&m=dev